+++ actix-session/CHANGES.md
@@ -3,7 +3,7 @@
 ## Unreleased
 
 - Add `redis-session-rustls` crate feature that enables `rustls`-secured Redis sessions.
-- Add `redis-pool` crate feature (off-by-default) which enables `RedisSessionStore::{new, builder}_pooled()` consrtuctors.
+- Add `redis-pool` crate feature (off-by-default) which enables `RedisSessionStore::{new, builder}_pooled()` constructors.
 - Rename `redis-rs-session` crate feature to `redis-session`.
 - Rename `redis-rs-tls-session` crate feature to `redis-session-native-tls`.
 - Remove `redis-actor-session` crate feature (and, therefore, the `actix-redis` based storage backend).